Timeout for all functions. Better selection of the slow functions.
